fix(app-control): make first-build verification actually winnable
two structural holes made every fresh app build fail or die:

1. register-before-launch: the launch check resolves through the runtime
   registry, but a new app was registered only by the PASS-path room bridge —
   circular, so a first build's launch check could never pass. the check now
   runs the same idempotent load-from-directory the bridge uses (hoisted to
   verification-helpers, one implementation, both callers) before launching.

2. cold-registry fail-soft: the create flow's optional "edit an existing
   app?" list read aborted at the 2s loopback deadline whenever the server's
   registry cache was cold (2.5s network fetch + workspace scans), surfacing
   as "The operation timed out." on the first build after every restart.
   the read now degrades to create-new (error-policy:J4); the load-bearing
   edit-path reads stay strict.

plugin suite 1528 pass, typecheck + lint clean.
